refactor: remove the wgproxy build tag; proxy compiles on all platforms
Proxy mode is a runtime decision (proxy.enabled) now, so the tag only
hid portable code. Merge proxy_enabled/proxy_disabled into an untagged
proxy_stack.go, drop -tags wgproxy from Makefile and CI, and let the
wgproxy tests join plain go test ./... everywhere.
